@@ -342,44 +342,52 @@
 </template>
 
 <script>
-//导入共通组件
 import Footer from '../components/Footer.vue';
+import { onMounted, onBeforeUnmount, nextTick, ref } from 'vue';
+import { useRouter } from 'vue-router';
 
 export default {
 	name: 'Index',
-	mounted() {
-		document.onscroll = () => {
-			//获取滚动条位置
-			let s1 = document.documentElement.scrollTop;
-			let s2 = document.body.scrollTop;
-			let scroll = s1 == 0 ? s2 : s1;
-			//获取视口宽度
-			let width = document.documentElement.clientWidth;
-
-			//获取顶部固定块
-			let search = this.$refs.fixedBox;
-
-			//判断滚动条超过视口宽度的12%时，搜索块变固定定位
-			if (scroll > width * 0.12) {
-				search.style.position = 'fixed';
-				search.style.left = '0';
-				search.style.top = '0';
-			} else {
-				search.style.position = 'static';
-			}
-		}
-	},
-	destroyed() {
-		//当切换到其他组件时，就不需要document滚动条事件，所以将此事件去掉
-		document.onscroll = null;
-	},
 	components: {
 		Footer
 	},
-	methods: {
-		toBusinessList(orderTypeId) {
-			this.$router.push({ path: '/businessList', query: { orderTypeId: orderTypeId } });
-		}
+	setup() {
+		const fixedBox = ref(null); // 使用 ref 来创建一个响应式的引用
+		const router = useRouter();
+
+		const handleScroll = () => {
+			let s1 = document.documentElement.scrollTop;
+			let s2 = document.body.scrollTop;
+			let scroll = s1 === 0 ? s2 : s1;
+			let width = document.documentElement.clientWidth;
+
+			if (scroll > width * 0.12) {
+				fixedBox.value.style.position = 'fixed';
+				fixedBox.value.style.left = '0';
+				fixedBox.value.style.top = '0';
+			} else {
+				fixedBox.value.style.position = 'static';
+			}
+		};
+
+		let removeScrollListener = () => {
+			document.onscroll = null;
+		};
+
+		onMounted(() => {
+			nextTick(() => {
+				document.onscroll = handleScroll;
+			});
+		});
+
+		onBeforeUnmount(removeScrollListener);
+
+		return {
+			fixedBox, // 将 fixedBox 包含在返回对象中，以便在模板中使用
+			toBusinessList: (orderTypeId) => {
+				router.push({ path: '/businessList', query: { orderTypeId: orderTypeId } });
+			}
+		};
 	}
 }
 </script>
