@@ -38,7 +38,7 @@
 	</div>
 </template>
 
-<script>
+<!-- <script>
 	import Footer from '../components/Footer.vue';
 
 	export default {
@@ -66,7 +66,7 @@
 					password: this.password
 				}).then(response => {
 					let user = response.data;
-
+					console.log(response);
 					if (user == null || user == '') {
 						alert('用户名或密码不正确！');
 					} else {
@@ -89,137 +89,200 @@
 			Footer
 		}
 	}
+</script> -->
+
+<script>
+import { ref } from 'vue';
+import Footer from '../components/Footer.vue';
+import { useRouter } from 'vue-router';
+import axios from 'axios';
+import { setSessionStorage } from '../common.js'; // 确保你有一个setSessionStorage方法
+
+export default {
+	name: 'Login',
+	components: {
+		Footer
+	},
+	setup() {
+		const router = useRouter();
+		const userId = ref('');
+		const password = ref('');
+
+		const login = () => {
+			if (userId.value === '') {
+				alert('手机号码不能为空！');
+				return;
+			}
+			if (password.value === '') {
+				alert('密码不能为空！');
+				return;
+			}
+
+			// 登录请求
+			axios.post('users/login', {
+				userId: userId.value,
+				password: password.value
+			}).then(response => {
+				const user = response.data;
+				console.log(response);
+				if (user == null || user == '') {
+					alert('用户名或密码不正确！');
+				} else {
+					// sessionStorage有容量限制，为了防止数据溢出，所以不将userImg数据放入session中
+					user.userImg = '';
+					setSessionStorage('user', user);
+					router.go(-1);
+				}
+			}).catch(error => {
+				console.error(error);
+			});
+		};
+
+		const register = () => {
+			router.push({
+				path: 'register'
+			});
+		};
+
+		return {
+			userId,
+			password,
+			login,
+			register
+		};
+	}
+}
 </script>
 
 <style scoped>
-	/****************** 总容器 ******************/
-	.wrapper {
-		width: 100%;
-		height: 100%;
-	}
+/****************** 总容器 ******************/
+.wrapper {
+	width: 100%;
+	height: 100%;
+}
 
-	/****************** header部分 ******************/
-	.wrapper header {
-		width: 100%;
-		height: 12vw;
-		background-color: #0097FF;
-		color: #fff;
-		font-size: 4.8vw;
+/****************** header部分 ******************/
+.wrapper header {
+	width: 100%;
+	height: 12vw;
+	background-color: #0097FF;
+	color: #fff;
+	font-size: 4.8vw;
 
-		position: fixed;
-		left: 0;
-		top: 0;
-		z-index: 1000;
+	position: fixed;
+	left: 0;
+	top: 0;
+	z-index: 1000;
 
-		display: flex;
-		justify-content: center;
-		align-items: center;
-	}
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
 
-	/****************** 表单部分 ******************/
-	.wrapper .form-box {
-		width: 100%;
-		margin-top: 12vw;
-	}
+/****************** 表单部分 ******************/
+.wrapper .form-box {
+	width: 100%;
+	margin-top: 12vw;
+}
 
-	.wrapper .form-box li {
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-		display: flex;
-		align-items: center;
-	}
+.wrapper .form-box li {
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+	display: flex;
+	align-items: center;
+}
 
-	.wrapper .form-box li .title {
-		flex: 0 0 18vw;
-		font-size: 3vw;
-		font-weight: 700;
-		color: #666;
-	}
+.wrapper .form-box li .title {
+	flex: 0 0 18vw;
+	font-size: 3vw;
+	font-weight: 700;
+	color: #666;
+}
 
-	.wrapper .form-box li .content {
-		flex: 1;
-	}
+.wrapper .form-box li .content {
+	flex: 1;
+}
 
-	.wrapper .form-box li .content input {
-		border: none;
-		outline: none;
-		width: 100%;
-		height: 4vw;
-		font-size: 3vw;
-	}
+.wrapper .form-box li .content input {
+	border: none;
+	outline: none;
+	width: 100%;
+	height: 4vw;
+	font-size: 3vw;
+}
 
-	.wrapper .button-login {
-		width: 100%;
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-	}
+.wrapper .button-login {
+	width: 100%;
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+}
 
-	.wrapper .button-login button {
-		width: 100%;
-		height: 10vw;
-		font-size: 3.8vw;
-		font-weight: 700;
-		color: #fff;
-		background-color: #38CA73;
-		border-radius: 4px;
+.wrapper .button-login button {
+	width: 100%;
+	height: 10vw;
+	font-size: 3.8vw;
+	font-weight: 700;
+	color: #fff;
+	background-color: #38CA73;
+	border-radius: 4px;
 
-		border: none;
-		outline: none;
-	}
+	border: none;
+	outline: none;
+}
 
-	.wrapper .button-register {
-		width: 100%;
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-	}
+.wrapper .button-register {
+	width: 100%;
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+}
 
-	.wrapper .button-register button {
-		width: 100%;
-		height: 10vw;
-		font-size: 3.8vw;
-		font-weight: 700;
-		/*与上面登陆按钮不同的只有颜色、背景色、边框不同*/
-		color: #666;
-		background-color: #EEE;
-		border: solid 1px #DDD;
-		border-radius: 4px;
+.wrapper .button-register button {
+	width: 100%;
+	height: 10vw;
+	font-size: 3.8vw;
+	font-weight: 700;
+	/*与上面登陆按钮不同的只有颜色、背景色、边框不同*/
+	color: #666;
+	background-color: #EEE;
+	border: solid 1px #DDD;
+	border-radius: 4px;
 
-		border: none;
-		outline: none;
-	}
+	border: none;
+	outline: none;
+}
 
-	/****************** 底部菜单部分 ******************/
-	.wrapper .footer {
-		width: 100%;
-		height: 14vw;
-		border-top: solid 1px #DDD;
-		background-color: #fff;
+/****************** 底部菜单部分 ******************/
+.wrapper .footer {
+	width: 100%;
+	height: 14vw;
+	border-top: solid 1px #DDD;
+	background-color: #fff;
 
-		position: fixed;
-		left: 0;
-		bottom: 0;
+	position: fixed;
+	left: 0;
+	bottom: 0;
 
-		display: flex;
-		justify-content: space-around;
-		align-items: center;
-	}
+	display: flex;
+	justify-content: space-around;
+	align-items: center;
+}
 
-	.wrapper .footer li {
-		display: flex;
-		flex-direction: column;
-		justify-content: center;
-		align-items: center;
+.wrapper .footer li {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
 
-		color: #999;
-		user-select: none;
-		cursor: pointer;
-	}
+	color: #999;
+	user-select: none;
+	cursor: pointer;
+}
 
-	.wrapper .footer li p {
-		font-size: 2.8vw;
-	}
+.wrapper .footer li p {
+	font-size: 2.8vw;
+}
 
-	.wrapper .footer li i {
-		font-size: 5vw;
-	}
+.wrapper .footer li i {
+	font-size: 5vw;
+}
 </style>
