@@ -62,7 +62,7 @@
 	</div>
 </template>
 
-<script>
+<!-- <script>
 	import Footer from '../components/Footer.vue';
 	import Backer from '../components/backer.vue';
 
@@ -147,101 +147,195 @@
 			Backer
 		}
 	}
+</script> -->
+
+<script>
+import { ref } from 'vue';
+import Footer from '../components/Footer.vue';
+import Backer from '../components/backer.vue';
+import { useRouter } from 'vue-router';
+import axios from 'axios'; // 确保已经安装并导入axios
+
+export default {
+	name: 'Register',
+	components: {
+		Footer,
+		Backer
+	},
+	setup() {
+		const router = useRouter();
+		const user = ref({
+			userId: '',
+			password: '',
+			userName: '',
+			userSex: 1
+		});
+		const confirmPassword = ref('');
+
+		const checkUserId = () => {
+			const phoneRegex = /^1[3-9]\d{9}$/;
+			if (user.value.userId === '') {
+				alert('手机号码不能为空！');
+				return;
+			}
+			// 验证手机号码是否符合规范
+			if (!phoneRegex.test(user.value.userId)) {
+				alert('请输入正确的手机号码！');
+				return;
+			}
+			axios.get(`users/${user.value.userId}`)
+				.then(response => {
+					if (response.data === 1) {
+						user.value.userId = '';
+						alert('此手机号码已存在！');
+					}
+				})
+				.catch(error => {
+					console.error(error);
+				});
+		};
+
+		const register = () => {
+			const phoneRegex = /^1[3-9]\d{9}$/;
+			if (user.value.userId === '') {
+				alert('手机号码不能为空！');
+				return;
+			}
+			if (!phoneRegex.test(user.value.userId)) {
+				alert('请输入正确的手机号码！');
+				return;
+			}
+			if (user.value.password === '') {
+				alert('密码不能为空！');
+				return;
+			}
+			if (user.value.password !== confirmPassword.value) {
+				alert('两次输入的密码不一致！');
+				return;
+			}
+			if (user.value.userName === '') {
+				alert('用户名不能为空！');
+				return;
+			}
+
+			// 注册请求
+			axios.post('users', user.value)
+				.then(response => {
+					if (response.data > 0) {
+						alert('注册成功！');
+						router.go(-1);
+					} else {
+						alert('注册失败！');
+					}
+				})
+				.catch(error => {
+					console.error(error);
+				});
+		};
+
+		return {
+			user,
+			confirmPassword,
+			checkUserId,
+			register
+		};
+	}
+}
 </script>
 
 <style scoped>
-	/****************** 总容器 ******************/
-	.wrapper {
-		width: 100%;
-		height: 100%;
-	}
+/****************** 总容器 ******************/
+.wrapper {
+	width: 100%;
+	height: 100%;
+}
 
-	/****************** header部分 ******************/
-	.wrapper header {
-		width: 100%;
-		height: 12vw;
-		background-color: #0097FF;
-		color: #fff;
-		font-size: 4.8vw;
+/****************** header部分 ******************/
+.wrapper header {
+	width: 100%;
+	height: 12vw;
+	background-color: #0097FF;
+	color: #fff;
+	font-size: 4.8vw;
 
-		position: fixed;
-		left: 0;
-		top: 0;
-		z-index: 1000;
+	position: fixed;
+	left: 0;
+	top: 0;
+	z-index: 1000;
 
-		display: flex;
-		justify-content: center;
-		align-items: center;
-	}
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
 
-	/****************** 表单部分 ******************/
-	.wrapper .form-box {
-		width: 100%;
-		margin-top: 12vw;
-	}
+/****************** 表单部分 ******************/
+.wrapper .form-box {
+	width: 100%;
+	margin-top: 12vw;
+}
 
-	.wrapper .form-box li {
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-		display: flex;
-		align-items: center;
-	}
+.wrapper .form-box li {
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+	display: flex;
+	align-items: center;
+}
 
-	.wrapper .form-box li .title {
-		flex: 0 0 18vw;
-		font-size: 3vw;
-		font-weight: 700;
-		color: #666;
-	}
+.wrapper .form-box li .title {
+	flex: 0 0 18vw;
+	font-size: 3vw;
+	font-weight: 700;
+	color: #666;
+}
 
-	.wrapper .form-box li .content {
-		flex: 1;
-	}
+.wrapper .form-box li .content {
+	flex: 1;
+}
 
-	.wrapper .form-box li .content input {
-		border: none;
-		outline: none;
-		width: 100%;
-		height: 4vw;
-		font-size: 3vw;
-	}
+.wrapper .form-box li .content input {
+	border: none;
+	outline: none;
+	width: 100%;
+	height: 4vw;
+	font-size: 3vw;
+}
 
-	.wrapper .button-login {
-		width: 100%;
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-	}
+.wrapper .button-login {
+	width: 100%;
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+}
 
-	.wrapper .button-login button {
-		width: 100%;
-		height: 10vw;
-		font-size: 3.8vw;
-		font-weight: 700;
-		color: #fff;
-		background-color: #38CA73;
-		border-radius: 4px;
+.wrapper .button-login button {
+	width: 100%;
+	height: 10vw;
+	font-size: 3.8vw;
+	font-weight: 700;
+	color: #fff;
+	background-color: #38CA73;
+	border-radius: 4px;
 
-		border: none;
-		outline: none;
-	}
+	border: none;
+	outline: none;
+}
 
-	.wrapper .button-register {
-		width: 100%;
-		box-sizing: border-box;
-		padding: 4vw 3vw 0 3vw;
-	}
+.wrapper .button-register {
+	width: 100%;
+	box-sizing: border-box;
+	padding: 4vw 3vw 0 3vw;
+}
 
-	.wrapper .button-register button {
-		width: 100%;
-		height: 10vw;
-		font-size: 3.8vw;
-		font-weight: 700;
-		color: #666;
-		background-color: #EEE;
-		border-radius: 4px;
+.wrapper .button-register button {
+	width: 100%;
+	height: 10vw;
+	font-size: 3.8vw;
+	font-weight: 700;
+	color: #666;
+	background-color: #EEE;
+	border-radius: 4px;
 
-		border: none;
-		outline: none;
-		border: solid 1px #DDD;
-	}
+	border: none;
+	outline: none;
+	border: solid 1px #DDD;
+}
 </style>
