@@ -1,16 +1,16 @@
 <template>
-    <div class="wrapper">
-        <header class="profile-header">
+	<div class="wrapper">
+		<header class="profile-header">
 			<div class="header-left">
 				<div class="avatar-frame">
-				    <img src="../assets/defaultphoto.png" alt="无法加载图片">
+					<img src="../assets/defaultphoto.png" alt="无法加载图片">
 				</div>
 				<h2 class="username">{{ user.userName }}</h2>
 			</div>
 			<div class="header-right">
 				<img @click="toindividual_inf" src="../assets/setting.png">
 			</div>
-        </header>
+		</header>
 		<ul class="money">
 			<li>
 				<img src="../assets/money.png">
@@ -48,53 +48,47 @@
 			</li>
 		</ul>
 
-        <Footer></Footer>
-    </div>
+		<Footer></Footer>
+	</div>
 </template>
 
 <script>
+import { useRouter } from 'vue-router';
 import Footer from '../components/Footer.vue';
+import { getSessionStorage } from '../common.js';
+
 export default {
-    data() {
-        return {
-            user: {}
-        }
-    },
-    created() {
-        this.user = this.$getSessionStorage('user');
-        // this.$axios.get(`users/${this.user}`)
-        //     .then(response => {
-        //         this.user = response.data;
-        //     }).catch(error => {
-        //         console.error(error);
-        //     });
-    },
-    methods: {
-        backfun() {
-            this.$router.push({ path: '/index' });
-        },
-        toindividual_inf() {
-            this.$router.push({ path: '/person' });
-        }
-    },
-    components: {
-        Footer
-    }
-}
+	components: {
+		Footer
+	},
+	setup() {
+		const router = useRouter(); // 使用 useRouter 获取路由实例
+		const user = getSessionStorage('user') || { userName: '未登录' }; // 获取用户数据
+
+		const toindividual_inf = () => {
+			router.push({ path: '/person' });
+		};
+
+		return {
+			user,
+			toindividual_inf
+		};
+	}
+};
 </script>
 
 <style scoped>
 .wrapper {
-    width: 100%;
-    height: 100%;
+	width: 100%;
+	height: 100%;
 	background-color: #f5f5f5;
-    /* display: flex;
+	/* display: flex;
     flex-direction: column;
     padding-bottom: 14vw; */
-    /* 为底部菜单留出空间 */
+	/* 为底部菜单留出空间 */
 }
 
-.wrapper header{
+.wrapper header {
 	width: 100%;
 	height: 20vw;
 	display: flex;
@@ -102,50 +96,50 @@ export default {
 	justify-content: space-between;
 }
 
-.wrapper header .header-left{
-    width: 80vw;
-    height: 20vw;
-    background-color: #f5f5f5;
-    /* text-align: center; */
-    display: flex;
-    align-items: center;
-    /* padding: 0 2vw; */
+.wrapper header .header-left {
+	width: 80vw;
+	height: 20vw;
+	background-color: #f5f5f5;
+	/* text-align: center; */
+	display: flex;
+	align-items: center;
+	/* padding: 0 2vw; */
 }
 
 .avatar-frame {
-    width: 12vw;
-    /* 使用视口宽度的10%作为宽度 */
-    height: 12vw;
-    /* 使用视口宽度的10%作为高度 */
-    border-radius: 50%;
-    /* overflow: hidden; */
-    /* border: 3px solid #ccc; */
-    margin-right: 2vw;
-    margin-left: 2vw;
+	width: 12vw;
+	/* 使用视口宽度的10%作为宽度 */
+	height: 12vw;
+	/* 使用视口宽度的10%作为高度 */
+	border-radius: 50%;
+	/* overflow: hidden; */
+	/* border: 3px solid #ccc; */
+	margin-right: 2vw;
+	margin-left: 2vw;
 }
 
 .avatar-frame img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
+	width: 100%;
+	height: 100%;
+	object-fit: cover;
 }
 
 .username {
-    font-size: 5vw;
-    /* 使用视口宽度的2%作为字体大小 */
-    color: #333;
-    /* margin-top: 5vw; */
-    /* 使用视口宽度的1%作为上边距 */
+	font-size: 5vw;
+	/* 使用视口宽度的2%作为字体大小 */
+	color: #333;
+	/* margin-top: 5vw; */
+	/* 使用视口宽度的1%作为上边距 */
 }
 
-.wrapper header .header-right{
+.wrapper header .header-right {
 	width: 20vw;
 	height: 20vw;
 	display: flex;
 	align-items: center;
 }
 
-.wrapper header .header-right img{
+.wrapper header .header-right img {
 	width: 7vw;
 	height: 7vw;
 	font-weight: 600;
@@ -153,32 +147,32 @@ export default {
 }
 
 .profile-menu {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    flex: 1;
-    /* 使菜单列表占据剩余的空间 */
+	list-style: none;
+	padding: 0;
+	margin: 0;
+	flex: 1;
+	/* 使菜单列表占据剩余的空间 */
 }
 
 .profile-menu-item {
-    padding: 8vw;
-    /* 使用视口宽度的2%作为内边距 */
-    font-size: 3vw;
-    /* 使用视口宽度的1.5%作为字体大小 */
-    color: #555;
-    border-bottom: 1px solid #e0e0e0;
-    cursor: pointer;
-    transition: background-color 0.3s;
+	padding: 8vw;
+	/* 使用视口宽度的2%作为内边距 */
+	font-size: 3vw;
+	/* 使用视口宽度的1.5%作为字体大小 */
+	color: #555;
+	border-bottom: 1px solid #e0e0e0;
+	cursor: pointer;
+	transition: background-color 0.3s;
 }
 
 .profile-menu-item:hover {
-    background-color: #e9e9e9;
+	background-color: #e9e9e9;
 }
 
-.wrapper .money{
+.wrapper .money {
 	width: 95%;
 	height: 30vw;
-	
+
 	display: flex;
 	justify-content: space-around;
 	align-items: center;
@@ -188,15 +182,15 @@ export default {
 	margin-right: 2vw;
 }
 
-.wrapper .money li{
+.wrapper .money li {
 	width: 30vw;
 	height: 20vw;
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
 	align-items: center;
-	
-/* 	user-select: none;
+
+	/* 	user-select: none;
 	cursor: pointer; */
 }
 
@@ -205,19 +199,19 @@ export default {
 	height: 6vw;
 }
 
-.wrapper .money li p{
+.wrapper .money li p {
 	margin-top: 1vw;
 	font-size: 3.2vw;
 }
 
-.wrapper .system-setting{
+.wrapper .system-setting {
 	width: 95%;
 	height: 30vw;
 	margin-top: 2vw;
-	
+
 	display: flex;
 	flex-wrap: wrap;
-	justify-content: space-between; 
+	justify-content: space-between;
 	align-items: center;
 	background-color: #fff;
 	border-radius: 5vw;
@@ -225,15 +219,15 @@ export default {
 	margin-right: 2.5vw;
 }
 
-.wrapper .system-setting li{
+.wrapper .system-setting li {
 	width: 22vw;
 	height: 15vw;
 	display: flex;
 	flex-direction: column;
 	justify-content: center;
 	align-items: center;
-	
-/* 	user-select: none;
+
+	/* 	user-select: none;
 	cursor: pointer; */
 }
 
@@ -242,10 +236,8 @@ export default {
 	height: 6vw;
 }
 
-.wrapper .system-setting li p{
+.wrapper .system-setting li p {
 	margin-top: 1vw;
 	font-size: 3.2vw;
 }
-
 </style>
-
